Fixing initialization problem and troubleshoting

+++ src/tools.cpp
@@ -51,15 +51,15 @@ MatrixXd Tools::CalculateJacobian(const VectorXd& x_state) {
     return Hj;
   }
 	//recover state parameters
-	float px = x_state(0);
-	float py = x_state(1);
-	float vx = x_state(2);
-	float vy = x_state(3);
+	double px = x_state(0);
+	double py = x_state(1);
+	double vx = x_state(2);
+	double vy = x_state(3);
 
 	//pre-compute a set of terms to avoid repeated calculation
-	float c1 = px*px+py*py;
-	float c2 = sqrt(c1);
-	float c3 = (c1*c2);
+	double c1 = px*px+py*py;
+	double c2 = sqrt(c1);
+	double c3 = (c1*c2);
 
 	//check division by zero
 	if(fabs(c1) < 0.0001){
